+++ spinetoolbox/helpers.py
@@ -25,6 +25,7 @@ import datetime
 import shutil
 import re
 import matplotlib
+import pathlib
 from PySide2.QtCore import Qt, Slot, QFile, QIODevice, QSize, QRect, QPoint, QUrl, QObject, QEvent
 from PySide2.QtCore import __version__ as qt_version
 from PySide2.QtCore import __version_info__ as qt_version_info
@@ -49,6 +50,7 @@ from PySide2.QtGui import (
     QPainter,
 )
 from spine_engine.utils.serialization import deserialize_path
+from spinedb_api.spine_io.gdx_utils import find_gams_directory
 from .config import DEFAULT_WORK_DIR, PLUGINS_PATH
 
 if os.name == "nt":
@@ -62,6 +64,11 @@ if _matplotlib_version[0] == 3 and _matplotlib_version[1] == 0:
     from pandas.plotting import register_matplotlib_converters
 
     register_matplotlib_converters()
+
+
+def home_dir():
+    """Returns user's home dir"""
+    return str(pathlib.Path.home())
 
 
 def format_log_message(msg_type, message, show_datetime=True):
@@ -756,6 +763,32 @@ class ChildCyclingKeyPressFilter(QObject):
         return QObject.eventFilter(self, obj, event)  # Pass event further
 
 
+def select_gams_executable(parent, line_edit):
+    """Opens file browser where user can select a Gams executable (i.e. gams.exe on Windows).
+
+    Args:
+        parent (QWidget, optional): Parent widget for the file dialog and message boxes
+        line_edit (QLineEdit): Line edit where the selected path will be inserted
+    """
+    start_dir = find_gams_directory()
+    if not start_dir:
+        start_dir = home_dir()
+    # noinspection PyCallByClass, PyTypeChecker, PyArgumentList
+    answer = QFileDialog.getOpenFileName(
+        parent, "Select GAMS Program (e.g. gams.exe on Windows)", start_dir
+    )
+    if answer[0] == "":  # Canceled (american-english), cancelled (british-english)
+        return
+    # Check that selected file at least starts with string 'gams'
+    _, selected_file = os.path.split(answer[0])
+    if not selected_file.lower().startswith("gams"):
+        msg = "Selected file <b>{0}</b> may not be a valid GAMS program".format(selected_file)
+        # noinspection PyCallByClass, PyArgumentList
+        QMessageBox.warning(parent, "Invalid GAMS Program", msg)
+        return
+    line_edit.setText(answer[0])
+
+
 def select_julia_executable(parent, line_edit):
     """Opens file browser where user can select a Julia executable (i.e. julia.exe on Windows).
     Used in SettingsWidget and KernelEditor.
@@ -766,21 +799,9 @@ def select_julia_executable(parent, line_edit):
     """
     # noinspection PyCallByClass, PyTypeChecker, PyArgumentList
     answer = QFileDialog.getOpenFileName(
-        parent, "Select Julia Executable (e.g. julia.exe on Windows)", os.path.abspath('C:\\')
+        parent, "Select Julia Executable (e.g. julia.exe on Windows)", home_dir()
     )
     if answer[0] == "":  # Canceled (american-english), cancelled (british-english)
-        return
-    # Check that it's not a directory
-    if os.path.isdir(answer[0]):
-        msg = "Please select a valid Julia Executable (file) and not a directory"
-        # noinspection PyCallByClass, PyArgumentList
-        QMessageBox.warning(parent, "Invalid Julia Executable", msg)
-        return
-    # Check that it's a file that actually exists
-    if not os.path.exists(answer[0]):
-        msg = "File {0} does not exist".format(answer[0])
-        # noinspection PyCallByClass, PyArgumentList
-        QMessageBox.warning(parent, "Invalid Julia Executable", msg)
         return
     # Check that selected file at least starts with string 'julia'
     _, selected_file = os.path.split(answer[0])
@@ -800,7 +821,7 @@ def select_julia_project(parent, line_edit):
         parent (QWidget, optional): Parent of QFileDialog
         line_edit (QLineEdit): Line edit where the selected path will be inserted
     """
-    answer = QFileDialog.getExistingDirectory(parent, "Select Julia project directory", os.path.abspath("C:\\"))
+    answer = QFileDialog.getExistingDirectory(parent, "Select Julia project directory", home_dir())
     if not answer:  # Canceled (american-english), cancelled (british-english)
         return
     line_edit.setText(answer)
@@ -816,15 +837,9 @@ def select_python_interpreter(parent, line_edit):
     """
     # noinspection PyCallByClass, PyTypeChecker, PyArgumentList
     answer = QFileDialog.getOpenFileName(
-        parent, "Select Python Interpreter (e.g. python.exe on Windows)", os.path.abspath("C:\\")
+        parent, "Select Python Interpreter (e.g. python.exe on Windows)", home_dir()
     )
     if answer[0] == "":  # Canceled
-        return
-    # Check that it's not a directory
-    if os.path.isdir(answer[0]):
-        msg = "Please select a valid Python interpreter (file) and not a directory"
-        # noinspection PyCallByClass, PyArgumentList
-        QMessageBox.warning(parent, "Invalid Python Interpreter", msg)
         return
     # Check that selected file at least starts with string 'python'
     _, selected_file = os.path.split(answer[0])
@@ -835,6 +850,29 @@ def select_python_interpreter(parent, line_edit):
         return
     line_edit.setText(answer[0])
     return
+
+
+def select_conda_executable(parent, line_edit):
+    """Opens file browser where user can select a conda executable.
+
+    Args:
+        parent (QWidget): Parent widget for the file dialog and message boxes
+        line_edit (QLineEdit): Line edit where the selected path will be inserted
+    """
+    # noinspection PyCallByClass, PyTypeChecker, PyArgumentList
+    answer = QFileDialog.getOpenFileName(
+        parent, "Select Conda Executable (e.g. conda.exe on Windows)", home_dir()
+    )
+    if answer[0] == "":  # Canceled
+        return
+    # Check that selected file at least starts with string 'conda'
+    _, selected_file = os.path.split(answer[0])
+    if not selected_file.lower().startswith("conda"):
+        msg = "Selected file <b>{0}</b> is not a valid Conda executable".format(selected_file)
+        # noinspection PyCallByClass, PyArgumentList
+        QMessageBox.warning(parent, "Invalid Conda selected", msg)
+        return
+    line_edit.setText(answer[0])
 
 
 def file_is_valid(parent, file_path, msgbox_title, extra_check=None):