+++ spinetoolbox/widgets/jupyter_console_widget.py
@@ -29,6 +29,8 @@ from spinetoolbox.widgets.project_item_drag import ProjectItemDragMixin
 from spinetoolbox.config import JUPYTER_KERNEL_TIME_TO_DEAD
 from spinetoolbox.widgets.kernel_editor import find_kernels
 from spinetoolbox.spine_engine_manager import make_engine_manager
+from spinetoolbox.widgets.settings_widget import resolve_conda_executable
+from spine_engine.execution_managers.conda_kernel_spec_manager import CondaKernelSpecManager
 
 # Set logging level for jupyter loggers
 traitlets_logger = logging.getLogger("traitlets")
@@ -139,7 +141,10 @@ class JupyterConsoleWidget(RichJupyterWidget):
         new_k_name_anchor = "<a style='color:#99CCFF;' title='{0}' href='#'>{1}</a>".format(k_path, self.kernel_name)
         self._toolbox.msg.emit(f"*** Starting {self.name()} (kernel {new_k_name_anchor}) ***")
         self._kernel_starting = True  # This flag is unset when a correct msg is received from iopub_channel
-        km = QtKernelManager(kernel_name=self.kernel_name)
+        km = QtKernelManager(kernel_name="conda-env-3-calliope-py")
+        conda_exe = self._toolbox.qsettings().value("appSettings/condaPath", defaultValue="")
+        conda_exe = resolve_conda_executable(conda_exe)
+        km.kernel_spec_manager = CondaKernelSpecManager(conda_exe=conda_exe)
         try:
             blackhole = open(os.devnull, 'w')
             cf = subprocess.CREATE_NO_WINDOW if sys.platform == "win32" else 0  # Don't show console when frozen