+++ spinetoolbox/widgets/settings_widget.py
@@ -17,13 +17,14 @@ Widget for controlling user settings.
 """
 
 import os
-from PySide2.QtWidgets import QWidget, QFileDialog, QMessageBox, QColorDialog
+from PySide2.QtWidgets import QWidget, QFileDialog, QColorDialog
 from PySide2.QtCore import Slot, Qt, QSize, QSettings
 from PySide2.QtGui import QPixmap
 from spine_engine.utils.helpers import (
     resolve_python_interpreter,
     resolve_julia_executable,
     resolve_gams_executable,
+    resolve_conda_executable,
     get_julia_env,
 )
 from .notification import Notification
@@ -39,11 +40,14 @@ from ..widgets.kernel_editor import (
     find_julia_kernels,
 )
 from ..helpers import (
+    select_gams_executable,
     select_python_interpreter,
     select_julia_executable,
     select_julia_project,
+    select_conda_executable,
     file_is_valid,
     dir_is_valid,
+    home_dir,
 )
 
 
@@ -67,6 +71,10 @@ class SettingsWidgetBase(QWidget):
         self._mouse_press_pos = None
         self._mouse_release_pos = None
         self._mouse_move_pos = None
+
+    @property
+    def qsettings(self):
+        return self._qsettings
 
     def connect_signals(self):
         """Connect signals."""
@@ -251,10 +259,11 @@ class SettingsWidget(SpineDBEditorSettingsMixin, SettingsWidgetBase):
     def connect_signals(self):
         """Connect signals."""
         super().connect_signals()
-        self.ui.toolButton_browse_gams.clicked.connect(self.browse_gams_path)
+        self.ui.toolButton_browse_gams.clicked.connect(self.browse_gams_button_clicked)
         self.ui.toolButton_browse_julia.clicked.connect(self.browse_julia_button_clicked)
         self.ui.toolButton_browse_julia_project.clicked.connect(self.browse_julia_project_button_clicked)
         self.ui.toolButton_browse_python.clicked.connect(self.browse_python_button_clicked)
+        self.ui.toolButton_browse_conda.clicked.connect(self.browse_conda_button_clicked)
         self.ui.pushButton_open_kernel_editor_python.clicked.connect(self.show_python_kernel_editor)
         self.ui.pushButton_open_kernel_editor_julia.clicked.connect(self.show_julia_kernel_editor)
         self.ui.toolButton_browse_work.clicked.connect(self.browse_work_path)
@@ -313,50 +322,29 @@ class SettingsWidget(SpineDBEditorSettingsMixin, SettingsWidgetBase):
             db_editor.ui.graphicsView.set_auto_expand_objects(checked)
 
     @Slot(bool)
-    def browse_gams_path(self, checked=False):
-        """Open file browser where user can select a GAMS program."""
-        # noinspection PyCallByClass, PyArgumentList
-        answer = QFileDialog.getOpenFileName(
-            self, "Select GAMS Program (e.g. gams.exe on Windows)", os.path.abspath("C:\\")
-        )
-        if answer[0] == "":  # Canceled (american-english), cancelled (british-english)
-            return
-        # Check that it's not a directory
-        if os.path.isdir(answer[0]):
-            msg = "Please select a valid GAMS program (file) and not a directory"
-            # noinspection PyCallByClass, PyArgumentList
-            QMessageBox.warning(self, "Invalid GAMS Program", msg)
-            return
-        # Check that it's a file that actually exists
-        if not os.path.exists(answer[0]):
-            msg = "File {0} does not exist".format(answer[0])
-            # noinspection PyCallByClass, PyArgumentList
-            QMessageBox.warning(self, "Invalid GAMS Program", msg)
-            return
-        # Check that selected file at least starts with string 'gams'
-        _, selected_file = os.path.split(answer[0])
-        if not selected_file.lower().startswith("gams"):
-            msg = "Selected file <b>{0}</b> may not be a valid GAMS program".format(selected_file)
-            # noinspection PyCallByClass, PyArgumentList
-            QMessageBox.warning(self, "Invalid GAMS Program", msg)
-            return
-        self.ui.lineEdit_gams_path.setText(answer[0])
-        return
+    def browse_gams_button_clicked(self, checked=False):
+        """Calls static method that shows a file browser for selecting a Gams executable."""
+        select_gams_executable(self, self.ui.lineEdit_gams_path)
 
     @Slot(bool)
     def browse_julia_button_clicked(self, checked=False):
-        """Calls static method that shows a file browser for selecting the Julia path."""
+        """Calls static method that shows a file browser for selecting a Julia path."""
         select_julia_executable(self, self.ui.lineEdit_julia_path)
 
     @Slot(bool)
     def browse_julia_project_button_clicked(self, checked=False):
-        """Calls static method that shows a file browser for selecting a Julia project."""
+        """Calls static method that shows a folder browser for selecting a Julia project."""
         select_julia_project(self, self.ui.lineEdit_julia_project_path)
 
     @Slot(bool)
     def browse_python_button_clicked(self, checked=False):
-        """Calls static method that shows a file browser for selecting Python interpreter."""
+        """Calls static method that shows a file browser for selecting a Python interpreter."""
         select_python_interpreter(self, self.ui.lineEdit_python_path)
+
+    @Slot(bool)
+    def browse_conda_button_clicked(self, checked=False):
+        """Calls static method that shows a file browser for selecting a Conda executable."""
+        select_conda_executable(self, self.ui.lineEdit_conda_path)
 
     @Slot(bool)
     def show_python_kernel_editor(self, checked=False):
@@ -430,7 +418,7 @@ class SettingsWidget(SpineDBEditorSettingsMixin, SettingsWidgetBase):
     def browse_work_path(self, checked=False):
         """Open file browser where user can select the path to wanted work directory."""
         # noinspection PyCallByClass, PyTypeChecker, PyArgumentList
-        answer = QFileDialog.getExistingDirectory(self, "Select Work Directory", os.path.abspath("C:\\"))
+        answer = QFileDialog.getExistingDirectory(self, "Select Work Directory", home_dir())
         if answer == '':  # Cancel button clicked
             return
         selected_path = os.path.abspath(answer)
@@ -509,6 +497,7 @@ class SettingsWidget(SpineDBEditorSettingsMixin, SettingsWidgetBase):
         use_python_kernel = int(self._qsettings.value("appSettings/usePythonKernel", defaultValue="0"))
         python_path = self._qsettings.value("appSettings/pythonPath", defaultValue="")
         python_kernel = self._qsettings.value("appSettings/pythonKernel", defaultValue="")
+        conda_path = self._qsettings.value("appSettings/condaPath", defaultValue="")
         work_dir = self._qsettings.value("appSettings/workDir", defaultValue="")
         save_spec = int(self._qsettings.value("appSettings/saveSpecBeforeClosing", defaultValue="1"))  # tri-state
         spec_show_undo = int(self._qsettings.value("appSettings/specShowUndo", defaultValue="2"))
@@ -572,6 +561,10 @@ class SettingsWidget(SpineDBEditorSettingsMixin, SettingsWidgetBase):
             self.ui.comboBox_python_kernel.setCurrentIndex(0)
         else:
             self.ui.comboBox_python_kernel.setCurrentIndex(ind)
+        conda_placeholder_txt = resolve_conda_executable("")
+        if conda_placeholder_txt:
+            self.ui.lineEdit_conda_path.setPlaceholderText(conda_placeholder_txt)
+        self.ui.lineEdit_conda_path.setText(conda_path)
         self.ui.lineEdit_work_dir.setText(work_dir)
         self.orig_work_dir = work_dir
         if save_spec == 0:
@@ -677,6 +670,9 @@ class SettingsWidget(SpineDBEditorSettingsMixin, SettingsWidgetBase):
             return False
         self._qsettings.setValue("appSettings/pythonPath", python_exe)
         self._qsettings.setValue("appSettings/pythonKernel", python_kernel)
+        # Conda
+        conda_exe = self.ui.lineEdit_conda_path.text().strip()
+        self._qsettings.setValue("appSettings/condaPath", conda_exe)
         # Work directory
         work_dir = self.ui.lineEdit_work_dir.text().strip()
         self.set_work_directory(work_dir)